extract Expr from js_ast
